+++ causalnerve/memory/__init__.py
@@ -2,10 +2,14 @@ from .structural_memory_bank import StructuralMemoryBank
 from .episodic_memory import EpisodicMemory
 from .motif_archive import MotifArchive
 from .recurrence_engine import RecurrenceEngine
+from .replay_engine import StructuralReplayEngine, GraphSnapshot, RevisionRecord
 
 __all__ = [
     "StructuralMemoryBank",
     "EpisodicMemory",
     "MotifArchive",
-    "RecurrenceEngine"
+    "RecurrenceEngine",
+    "StructuralReplayEngine",
+    "GraphSnapshot",
+    "RevisionRecord"
 ]